refactor: replace auth controller
